+++ tests/unit.test.js
@@ -1,0 +1,20 @@
+// @ts-nocheck
+/* eslint-disable no-undef */
+/* eslint-disable no-debugger */
+/* eslint-disable no-unused-vars */
+/* cSpell:disable */
+
+import main from "../index.js";
+import { createRequire } from "module";
+const require = createRequire(import.meta.url);
+const timeout = 60000;
+
+describe('do tests work?', () => {
+	test('a = a', () => {
+		expect(true).toBe(true);
+	});
+});
+
+
+//todo unit tests...
+
